requirements.txt
pyjwt
requests
pyyaml
flask
slackclient
python-dateutil
schedule

